+++ src/lib_shell/shell_metrics.ml
@@ -346,22 +346,22 @@ end
 
 module Proto_plugin = struct
   module type PROTOMETRICS = sig
-    type t = {cycle : float; consumed_gas : float}
-
     val hash : Protocol_hash.t
 
-    val decode_metadata : bytes -> t
+    val update_metrics :
+      protocol_metadata:bytes ->
+      Fitness.t ->
+      (cycle:float -> consumed_gas:float -> round:float -> unit) ->
+      unit Lwt.t
   end
 
   module UndefinedProtoMetrics (P : sig
     val hash : Protocol_hash.t
   end) =
   struct
-    type t = {cycle : float; consumed_gas : float}
-
     let hash = P.hash
 
-    let decode_metadata _ = {cycle = -1.; consumed_gas = -1.}
+    let update_metrics ~protocol_metadata:_ _ _ = Lwt.return_unit
   end
 
   let proto_metrics_table : (module PROTOMETRICS) Protocol_hash.Table.t =
@@ -391,6 +391,7 @@ module Chain_validator = struct
     ignored_head_count : Prometheus.Counter.t;
     branch_switch_count : Prometheus.Counter.t;
     head_increment_count : Prometheus.Counter.t;
+    head_round : Prometheus.Gauge.t;
     validation_worker_metrics : Worker.t;
     head_cycle : Prometheus.Gauge.t;
     consumed_gas : Prometheus.Gauge.t;
@@ -459,6 +460,16 @@ module Chain_validator = struct
         ?subsystem
         "consumed_gas"
     in
+    let head_round =
+      let help = "Current Round" in
+      Prometheus.Gauge.v_label
+        ~label_name
+        ~help
+        ~namespace
+        ?subsystem
+        "head_round"
+    in
+
     let validation_worker_metrics =
       Worker.declare ~label_names:[label_name] ~namespace ?subsystem ()
     in
@@ -469,10 +480,16 @@ module Chain_validator = struct
         ignored_head_count = ignored_head_count label;
         branch_switch_count = branch_switch_count label;
         head_increment_count = head_increment_count label;
+        head_round = head_round label;
         validation_worker_metrics = validation_worker_metrics [label];
         head_cycle = head_cycle label;
         consumed_gas = consumed_gas label;
       }
+
+  let update_proto_metrics_callback ~metrics ~cycle ~consumed_gas ~round =
+    Prometheus.Gauge.set metrics.head_cycle cycle ;
+    Prometheus.Gauge.set metrics.consumed_gas consumed_gas ;
+    Prometheus.Gauge.set metrics.head_round round
 
   let update_ref = ref (fun () -> Lwt.return_unit)