+++ src/lib_shell/shell_metrics.mli
@@ -79,11 +79,13 @@ module Proto_plugin : sig
    * allows to decode protocol data payload and provide back basic
    * types that can be used as metrics. *)
   module type PROTOMETRICS = sig
-    type t = {cycle : float; consumed_gas : float}
-
     val hash : Protocol_hash.t
 
-    val decode_metadata : bytes -> t
+    val update_metrics :
+      protocol_metadata:bytes ->
+      Fitness.t ->
+      (cycle:float -> consumed_gas:float -> round:float -> unit) ->
+      unit Lwt.t
   end
 
   (** Emtpy metrics module. All metrics are -1. *)
@@ -107,12 +109,16 @@ module Chain_validator : sig
     ignored_head_count : Prometheus.Counter.t;
     branch_switch_count : Prometheus.Counter.t;
     head_increment_count : Prometheus.Counter.t;
+    head_round : Prometheus.Gauge.t;
     validation_worker_metrics : Worker.t;
     head_cycle : Prometheus.Gauge.t;
     consumed_gas : Prometheus.Gauge.t;
   }
 
   val init : string trace -> Chain_id.t -> t
+
+  val update_proto_metrics_callback :
+    metrics:t -> cycle:float -> consumed_gas:float -> round:float -> unit
 
   val update_proto : (unit -> unit Lwt.t) -> unit
 end