+++ CHANGES.rst
@@ -31,7 +31,8 @@ Node
 - Added a block validator metric to expose the number of operation per
   pass for each new block validated.
 
-- Added a protocol specific metrics, head_cycle and head_consumed_gas.
+- Added a protocol specific metrics, head_cycle, head_consumed_gas and
+  head_round.
 
 Client
 ------